test: avoid starting two servers concurrently in a test
