+++ playground/css/__tests__/postcss-plugins-different-dir/postcss-plugins-different-dir.spec.ts
@@ -6,7 +6,7 @@ import { getBgColor, getColor, page, ports } from '~utils'
 test('postcss plugins in different dir', async () => {
   const port = ports['css/postcss-plugins-different-dir']
   const server = await createServer({
-    root: path.join(__dirname, '..', '..', 'tailwind'),
+    root: path.join(__dirname, '..', '..', '..', 'tailwind'),
     logLevel: 'silent',
     server: {
       port,
