+++ playground/css/postcss-caching/serve.ts
@@ -1,0 +1,10 @@
+// this is automatically detected by playground/vitestSetup.ts and will replace
+// the default e2e test serve behavior
+
+// The server is started in the test, so we need to have a custom serve
+// function or a default server will be created
+export async function serve(): Promise<{ close(): Promise<void> }> {
+  return {
+    close: () => Promise.resolve()
+  }
+}
