+++ playground/js-sourcemap/__tests__/js-sourcemap.spec.ts
@@ -44,3 +44,9 @@ if (!isBuild) {
     expect(true).toBe(true)
   })
 }
+
+test.runIf(isBuild)('should not output sourcemap warning (#4939)', () => {
+  serverLogs.forEach((log) => {
+    expect(log).not.toMatch('Sourcemap is likely to be incorrect')
+  })
+})
