+++ playground/tailwind-sourcemap/postcss.config.js
@@ -1,3 +1,5 @@
+console.log(__dirname + '/tailwind.config.js')
+
 module.exports = {
   plugins: {
     tailwindcss: { config: __dirname + '/tailwind.config.js' }
