--- playground/vue-sourcemap/__tests__/build.spec.ts
@@ -1,7 +1,0 @@
-import { isBuild, serverLogs } from '~utils'
-
-test.runIf(isBuild)('should not output sourcemap warning (#4939)', () => {
-  serverLogs.forEach((log) => {
-    expect(log).not.toMatch('Sourcemap is likely to be incorrect')
-  })
-})
