+++ playground/vue-sourcemap/__tests__/vue-sourcemap.spec.ts
@@ -2,8 +2,10 @@ import { URL } from 'node:url'
 import {
   extractSourcemap,
   formatSourcemapForSnapshot,
+  isBuild,
   isServe,
-  page
+  page,
+  serverLogs
 } from '~utils'
 
 describe.runIf(isServe)('serve:vue-sourcemap', () => {
@@ -96,5 +98,11 @@ describe.runIf(isServe)('serve:vue-sourcemap', () => {
     const js = await res.text()
     const map = extractSourcemap(js)
     expect(formatSourcemapForSnapshot(map)).toMatchSnapshot()
+  })
+})
+
+test.runIf(isBuild)('should not output sourcemap warning (#4939)', () => {
+  serverLogs.forEach((log) => {
+    expect(log).not.toMatch('Sourcemap is likely to be incorrect')
   })
 })
